Handle map/unmap events on Wayland's wxGLCanvasEGL
Fixes hiding a wxGLCanvas on Wayland, either directly (`->Show(false)`)
or indirectly (e.g. when it is contained in a wxNotebook).

On Wayland, unlike on X11, the window and surface that we create with
`wl_egl_window_create` and `eglCreatePlatformWindowSurface` is detached
from the GTK widget associated to the canvas, thus it is not
automatically mapped or unmapped when the associated GTK widget is.
Rather, we need to manually synchronize it with the widget's state.

As Wayland/EGL does not provide any way to temporally hide the window,
we need to fully destroy it on unmap, and re-create it on map.

Fixes #22580