Deal with Wayland window surface mapping on wxGLCanvasEGL
This deals with a very inconvenient behaviour of GTK/GDK with respect
to the Wayland surfaces returned by gdk_wayland_window_get_wl_surface.

On Wayland, the first time we create a wxFrame, place a wxGLCanvasEGL
inside it, and show it, everything works: On the map handler for
wxGLCanvasEGL's GTK widget, gdk_wayland_window_get_wl_surface gives us
a wl_surface which we can set up our canvas on.

However, if we the hide the wxFrame, the re-display it, then on the map
handler for wxGLCanvasEGL's GTK widget, we will get a nullptr when
calling gdk_wayland_window_get_wl_surface.

What's going on inside GTK/GDK is that when the GdkWindow for the
top-level window is initially created, the underlying wl_surface is
created immediately. When it is hidden, that wl_surface is destroyed.
And when it is re-shown, that wl_surface is created again only at the
very last minute, when the top-level window is mapped again, and well
after the wxGLCanvasEGL's GTK widget map handler has been called.

Side note: This may not have been intended, and has been changed for
GTK4, see: https://gitlab.gnome.org/GNOME/gtk/-/commit/5d3cec5441ca

One way to deal with this is to use the map-event and unmap-event
signals, which happen later than the map and unmap signals, However,
this means that the wxGLCanvasEGL is not setup when gtk_widget_show()
returns. This breaks some assumptions that wxWidgets currently depends
on (and which work well on X11).

The other way to deal with this is to *also* connect to the toplevel's
window map signal. There gdk_wayland_window_get_wl_surface no longer
returns nullptr, so we can finish setting up the canvas.
This is a bit of a pain since the canvas can be re-parented to another
top-level window, so we also need to track toplevel window changes to
re-associate our event handlers, but it works.